API 서버 선택 UI 추가
