+++ src/components/Logo/Logo.js
@@ -19,18 +19,6 @@ class Logo extends Component {
         }, 300);
     };
 
-    toggleFixed = () => {
-        if (this.props.control.Controller === "fixed") {
-            this.props.setControllerState("min");
-        } else {
-            this.props.setControllerState("fixed");
-        }
-
-        setTimeout(() => {
-            window.dispatchEvent(new Event('resize'));
-        }, 300);
-    };
-
     render() {
 
         let instanceParam = (this.props.objects && this.props.objects.length > 0) ? "?objects=" + this.props.objects.map((d) => {
@@ -50,12 +38,6 @@ class Logo extends Component {
                 </div>
                 <div className="side-icon">
                     <div>
-                        <div className="fixed-wrapper" onClick={this.toggleFixed}>
-                            <span className={"fixed-btn " + this.props.control.Controller}>
-                                <i className="fa fa-heart-o no-fixed" aria-hidden="true"></i>
-                                <i className="fa fa-heart fixed" aria-hidden="true"></i>
-                            </span>
-                        </div>
                         <div className="max-wrapper" onClick={this.toggleMinMax}>
                             <span className={"max-btn " + this.props.control.Controller}>
                                 <i className="fa fa-angle-left max" aria-hidden="true"></i>
