+++ src/components/Menu/Menu.js
@@ -72,7 +72,7 @@ class Menu extends Component {
         return (
             <div className="menu-div">
                 <div className="menu">
-                    <div className="logo">
+                    {/*<div className="logo">
                         <NavLink to={"/" + instanceParam} activeClassName="active" onClick={this.menuClick.bind(this, "/")}>
                             <div className="logo-icon"><i className="fa fa-bolt" aria-hidden="true"></i></div>
                             <div className="logo-text-div">
@@ -80,7 +80,7 @@ class Menu extends Component {
                                 <div className="logo-text logo-text-2">PAPER</div>
                             </div>
                         </NavLink>
-                    </div>
+                    </div>*/}
                     <NavLink className={"menu-item " + (this.state.menu === "/topology" ? "active" : "")} to={"/topology"  + instanceParam} onClick={this.menuClick.bind(this, "/topology")}>
                         <div>
                             <div className="icon"><i className="fa fa-braille" aria-hidden="true"></i></div>
