+++ src/components/Navigator/Navigator.css
@@ -4,6 +4,7 @@
     font-family: 'Bungee', sans-serif;
     background-color: rgba(255, 255, 255, 0.2);
     background-color: #191919;
+    background-color: rgba(0,0,0,0.6);
     height: 50px;
     box-sizing: border-box;
 }
