+++ src/components/Paper/Paper.css
@@ -1,8 +1,8 @@
 .papers {
-
-    margin: 0 auto;
+    padding-left: 40px;
     height: calc(100% - 61px);
     position: relative;
+
 }
 
 .papers > div {
